Allow arbitrary partition structure for sinks
Also add support for partitions to be specified at config time
(the common case), as opposed to deriving from generated records.
HiveSupport split out to new file and reusable Partitions object
pulled up.
