+++ examples/src/main/scala/commbank/coppersmith/examples/Pivot.scala
@@ -11,7 +11,7 @@ import commbank.coppersmith.api.scalding._
 import commbank.coppersmith.examples.thrift.Movie
 
 case class PivotFeaturesConfig(conf: Config) extends FeatureJobConfig[Movie] {
-  val partitions     = ScaldingDataSource.Partitions.unpartitioned
+  val partitions     = Partitions.unpartitioned
   val movies         = HiveTextSource[Movie, Nothing](new Path("data/movies"), partitions)
 
   val featureSink    = EavtSink.configure("userguide", new Path("dev"), "movies")
